volume/app/requirements.txt
waitress==2.0.0
matplotlib==3.3.4
Flask==1.1.2
tensorflow==2.5.0
numpy==1.20.2
requests==2.25.1
mysql_connector_repackaged==0.3.1
Pillow==8.3.1
